test using semantic model
