+++ src/Compilers/CSharp/Test/Semantic/CSharpCompilerSemanticTest.csproj
@@ -137,6 +137,7 @@
     <Compile Include="Semantics\SyntaxTreeRootTests.cs" />
     <Compile Include="Semantics\SynthesizedStaticConstructorTests.cs" />
     <Compile Include="Semantics\TryCatchTests.cs" />
+    <Compile Include="Semantics\TypeOfTests.cs" />
     <Compile Include="Semantics\UnsafeTests.cs" />
     <Compile Include="Semantics\UserDefinedConversionTests.cs" />
     <Compile Include="Semantics\UseSiteErrorTests.cs" />
