+++ src/Compilers/CSharp/Test/Semantic/Semantics/TypeOfTests.cs
@@ -1,0 +1,36 @@
+﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.CSharp.Test.Utilities;
+using Roslyn.Test.Utilities;
+using Xunit;
+using System.Linq;
+
+namespace Microsoft.CodeAnalysis.CSharp.UnitTests
+{
+    public class TypeofTests : CSharpTestBase
+    {
+        [Fact, WorkItem(1720, "https://github.com/dotnet/roslyn/issues/1720")]
+        public void GetSymbolsOnResultOfNameof()
+        {
+            var source = @"
+class C
+{
+    public C(int i)
+    {
+        typeof(C).GetField("" "").SetValue(null, new C(0));
+    }
+}
+";
+            var compilation = CreateCompilationWithMscorlib45(source);
+            var tree = compilation.SyntaxTrees[0];
+            var model = compilation.GetSemanticModel(tree);
+            var node = (ObjectCreationExpressionSyntax) tree.GetRoot().DescendantNodes().Where(n => n.ToString() == "new C(0)").Last();
+            var identifierName = node.Type;
+
+            var symbolInfo = model.GetSymbolInfo(node);
+            Assert.Equal(symbolInfo.Symbol.ToTestDisplayString(), "C..ctor(System.Int32 i)");
+
+            Assert.Equal(TypeKind.Class, model.GetTypeInfo(node).Type.TypeKind);
+
+        }
+    }
+}
